replace all ord/chr
ord could always be dropped because it was used on bytes being iterated
over, or indexed bytes.

chr gets replaced with bytes([]) because that form is considered to be
most readable.

this is the last step that could be easily made to support python2 --
just add a `bytes` function that behaves like python3's `bytes` when
passed an iterable, and an `ord` function that is a lambda x:x in
python3 (because slicing a character already does the ord) if the
version is < 3.