respect NSTART=1, enqueue messages sent while an exchange is active
this introduces a backlog mapping remotes to lists of enqueued packages.

the sending function was split into an enqueueForSending function (which
may or may not send a message immediately), and a send function, which
gets called whenever the message is really sent. the creation of an
exchange, if required, is delayed into the send functon, for the timeout
only starts to be meaningful after the package is sent.