introduce objects representing a client side observation
this incurs renaming of the old Observation class to ServerObservation,
but the interfaces for applications don't change (callback/args/kwargs
passed into the request function get registered with the
ClientObervation for callback when it is created).

this allows catching errors like server-side cancelled observations, and
observations that are not accepted.

this partially reverts 4fdabc8a5 -- the interface in request is kept,
but the callback/args/kwargs data is no stored in the cbs tuple.