indicate that the request has been transmitted correctly
without this, the tests would spew warnings, as even in block2 phase
incoming messages would be routed through process_block1_in_response
(with `Block1 option completely ignored by server, assuming it knows
what it is doing.` showing as a warning).

it was forgotten to set the _request_transmitted_completely flag when
first introduced.