adapt test case
assigning a string to tuple interface of .opt.uri_path raised an
AttributeError in older versions, and the test case was left broken to
show that this needs a decision taken.

the decision is now to accept that strings are iterable in python, and
that whoever passes a string in somewhere where an iterable is expected
is on its own, as he usually is in python. (the special case of
basestring being not a typical iterator is not special enough to warrant
implementing additional safeguards). if you disagree, feel free to
convice me otherwise.