introduce unresolved_remote in message
this is a way for a message creator to specify the message's destination
in an intermediate way: not as semantically correct as setting the
Uri-Host option (which wouldn't allow implementing proxying), and not as
low-level as setting the `remote` property (which is difficult to get
right anyway).

as a result, the ProxyForwarder could be drastically simplified. note
that its constructor api changed, as host:port splitting is now handled
by the protocol.