rework of resources
the original txThings resources were modelled after twisted concepts i
am not familiar with and that were not flexible enough in the asyncio
port when it came to where the application can do async stuff.

this unifies the resource and site concepts -- the server context just
gets a resource that will have its .render(request) method called. if
that resource is a 'Site' that dispatches the request to other
resources, that's fine with the context but none of its concern.

as resources are not necessarily persistent any more in this concept,
the way resources interact with persistence was changed here.