request creation: step towards full-featured .request() interface
this allows adding an exchange_monitor to a request at request creation
using .request(), paving the way towards a deprecation of manually
creating a Request object on a context.

(that gave more flexibility w/rt arguments, but prove difficult to work
with when the proxy context implemented things completely different).

also, the proxy context now supports that interface as well.