protocol responders: memorize own key
assembling responders critically depends on the Responders to be matched
by a constant key. remembering that key inside the object is a step
towards extracting all data the protocol needs early enough to allow
applications to mutate the message.

strictly speaking, the NameBasedVirtualHost was violating the api by
modifying the uri_path attribute, but we'll let that slip as a step
towards mutable requests.