proxy: rework async main to allow more control
this allows the test routines to really shut down the proxy servers,
which previously led to spurious failures in the test suite due to
lingering servers. (REUSEADDR doesn't really help there either).

as a result, the cli proxy's main is now not a coroutine any more, but
an object for whose .initialized one can wait.