abandon proxy.server.ProxiedResource
the ProxiedResource became a thinner and thinner object that only
delegates the process of obtaining a result for a response to the proxy
server itself.

by making the proxy server a resource (which would typically but not
necessarily be used as the main site of a server context),
customizations by subclassing are possible without adding additional
hooks to an object that is devoid of own purpose.

this commit modifications in all proxy server implementations, see the
diff on aiocoap/cli/proxy.py on what needs to be done.