observation proxying: delay destruction of last response
the information on the observation from the client being requested
arrives quite early, before the response is rendered. immediate action
on it led to destruction of the data that would have to be returned.

this delays the cleanup; as the response matching process is
synchronous, it is safe to only delay it until the async loop is run
again.