proxy observations: ClientObservation callbacks
while originally a callback was registered for each ServerObservation,
this did not provide a clean way to handle errors, and would in future
also have caused difficulties as callbacks can't be unregistered from
ClientObservation objects.

instead, the callbacks are now central to the proxy, and dispatched to
all users; errors now only show up in the logs if the ClientObservations
do not respond to them by deregistering (which happens anyway in the
typical cancellation scenarios, eg. a non-successful response from the
server).