Request.handle_response: avoid raising asyncio.futures.InvalidStateError
previously, situations like ResourceChanged flying out of message
assembly were caught and turned into an exception on the request, but
the protocol code handling the incoming message had an
asyncio.futures.InvalidStateError raised later.

as a result, the _process_response function would not terminate, and if
the response message came in a CON, neither ACK nor RST would be sent.