Make sure observation is not GC'd when it has callbacks
The previous fix for [12] was incorrect in that it required the
request.observation to be kept around, while actualy the object is not
required for the observation to work any more once there are callbacks
in place.

This makes the observation non-weak as soon as callbacks are there. It
also shows a flaw in the unit test (there are registered callbacks, so
the GCing and the accompanying warning war incorrect), which will be
fixed next.

[12]: https://github.com/chrysn/aiocoap/issues/12