error: More and consistent code errors, deprecate NoResource
Exception classes for CoAP error codes are now consistently named after
their code names. Exceptions that actually distinguish behaviors within
a code (eg. "method not found in types list" vs "method not implemented
by this resource") are kept distinct, but the old NoResource exception
did not have a reason not to be NotFound and is thus deprecated.