OSCOAP plugtest: Implement self-service interface on server
The server is now not controlled from the command line any more, but
can be controlled via the /sequence-numbers resource.

The auto-exit feature is dropped (which means that new users need to
manually reset the server instead of the operator running it in a test 1
auto-exit loop. The auto-increment feature is implicit anyway since the
testnumber-independent server resource implementation.