UDP6 transport: Don't expose V4MAPPED addresses in URIs
The UDP6EndpointAddress.uri property used to return addresses like
'coap://[::ffff:127.0.0.1]' (ie. containing a V4MAPPED IPv6 address),
which is not understandable outside a hybrid stack. Now, those
addresses are detected and expressed in IPv4 notation, which is just as
usable in a V6 stack.

This especially affected the resource directory.