Rename UnsupportedMediaType to UnsupportedContentFormat
The error.UnsupportedMediaType exception and
codes.UNSUPPORTED_MEDIA_TYPE value have been renamed to
error.UnsupportedContentFormat and codes.UNSUPPORTED_CONTENT_FORMAT
respectively to reflect the correct names as per RFC7252. The old names
are still available for compatibility, but are considered deprecated and
will be removed in a future version.

News-Section: Deprecations