OSCOAP Plugtest: Adapt test 16
The sequence numbers passed around between protect and unprotect
functions used to be shortened partial IVs, but now they are full-length
partial IVs, and just appending a byte would have thrown off the
XOR-ing.

This function just as surely produces a modified partial IV, and that
over the full range of sequence numbers.