Rework PathCapable subsites: behave like root
This is a subtle but for subsite applications breaking change in how
subsites are handled. Subsites now receive messages with Uri-Path
options just as they would appear if the subsite was a root site and the
previous path options were not sent, ie. if a subsite is added as

>>> site.add_resource(['batch'], subsite)

then it will receive a request to </subsite/foo> as a render request
with Uri-Path ('foo',) as before, but a request to </subsite/> with
Uri-Path (), as it would have if it were the root site and the request
went to </foo> and </> respectively.

Requests with an Uri-Path ("",), which would previously be sent on
</subsite/> requests, now only occur on </subsite//> requests (are they
even legal?). The registration of any root resource in a subsite needs
to be updated accordingly.

This restores consistency between sites and subsites, and allows the
presence of a root resource in the root site in the first place.

Applications that both want to export a </x> resource and have a subsite
directly under it (that's typically those that used the .rt and .if_
attributes of a Site) now need to add distinct resources (see updated
Site documentation), but are asked reconsider their URI design.

Closes: https://github.com/chrysn/aiocoap/issues/63

News-Section: Resources