client observations: report errors when initial request failed
The ClientObservation's callback registration (and thus the asynchronous
iteration interface) used not to report errors that happened before the
setup of the callbacks.

This resulted in the earlier convention that the success of the
observation must be checked in the result. This is now not required any
more; the observation will reliably report at least an error.