OSCOAP: (request_kid, request_partiv) rework
Let's start with stating that observation results previously used the
wrong partial IV in their external AAD -- should have been the original
request's, but was their own sequence number.

This led to a general review of where the data of the request_partiv was
used. The second input/output argument of (un)protect is now used to
transport both request_kid and request_partiv in one go, and the
(un)protect workflow was generally reworked to

* accept an explicit sequence number at any time
* factor out the partial_iv_short padding into a function
* drop some of the hex/int-parsing in favor of int.{from,to}_hex