protocol: Move blockwise into own Response class
Blockwise requests will, after the next commits, not be handled by the
same Request object that deals with timeouts and response error
handling, but by a dedicated BlockwiseRequest class that instruments the
underlying requester (typically a Protocol) to perform the individual
exchanges of the blockwise operation.

This allows reusing the blockwise mechanism in OSCOAP's inner-blockwise,
and should make it easier to make observation of large resources behave
correctly.

Most applications will not need to be changed, but if an application
constructs requests not using the Protocol's .request method but by
instanciating a Request directly, it now needs to instanciate a
BlockwiseRequest or a Request rather than instanciating a Request and
passing handle_blockwise=True/False.

News-Section: Breaking changes