proxy.client: Fix .observation -> None change
Introduced in aaf2b838, Requests now always have an .observation
property rather (which is None for requests that are not observations).
If applications previously checked for presence of .observation, they
now need to check wheter it is None or not.

(This is announced to News here where I found it rather than in aaf2b838
that caused it; I though about reverting to the old behavior and
checking for presence in protocol.py too, but this seems cleaner until
there is a more generic "has single or multiple answers" mechanism in
place).

News-Section: Breaking changes