message: Manual copy rather than deepcopy
The Message.copy() method (and other copying methods like the block
dissections that so far did not use the newer .copy() method) used to do
a plain deepcopy. This doesn't go well with transports (upcoming)
transports that use socket objects as their addresses, and generally
wasn't as lightweight as .copy() was meant to be.