security, proxy: Only log requests when configured
The aiocoap-proxy command line tool used to unconditionally log its
traffic to `/tmp/proxy-{out,in}.log`; now, those locations are
configurable and default to not logging.

As this might result in unaware logging of traffic, it is considered a
security issue. Users of the proxy tool are advised to set appropriate
security permissions on the resulting files, and to upgrade as soon as
their stability requirements permit.

News-Section: Security