tests: use hostnames as discovered
Which hostnames are available for ::1, 127.0.0.1 or both depends on the
platform (observed differences between Debian and Termux); discovering
the behavior of names at test time seems to be the right thing to do.

Names that just happened to be 'localhost' for testing purposes were
renamed to avoid showing up on reviews where that string should be
parameterized.