interfaces: First specification of ".remote interface"
So far, Message.remote objects were pretty free-form and had only
informal interfaces. This codifies the established interface in an ABC
and cleans it up a little (.port was only used in unit tests and more of
a remnant from back when .remote was used by all parties as a socket
address tuple).