Move token and MID handling away from global context
This is the first step in a restructuring operation that is required to
have both MID-based transports (UDP/DTLS), token-only transports (TCP)
and token-independent transports (OSCORE) all in one context.

This state represents an intermediate step that again passes the test
suite; don't expect all mechanisms introduced here to stay. (Especially
the ResponseProxy and its observation are awkward, but provide the
current interface to stay compatible with the test suite and show that
nothing relevant has been lost on the way until here).