tests: Adapt test_notreally
That test used to test a particular behavior that was not really
well motivated (just considered a nice thing to have back then), that
is, allowing a resource to declare its intent to be obsevable, and then
say it isn't while rendering the resource and sending out the
representation without observability already.

The test was adapted to cover the code paths of the now deprecatd
underlying API, and will probably go away when that API does.