Interfaces: Implement message recognition
This allows intefaces to see that a message already has a remote set,
and to tell the dispatch that it can take care of it.

This is especially important given that as announced in dc0dbfa2,
block-wise transfer often requires continuous access now, and simple6
could not send consistent source addresses (ie. use sockets) even when
the block handling set the remote appropriately.