servers: Accept None as port to be filled by transport
With the introduction of TLS server, and while all kinds of transports
are automatically set up by create_server_context, the default value of
('::, COAP_PORT) is no longer convenient for the binding specification.

All server-side host/port combinations now default to a port of None,
which can also be explicitly set (eg. `bind=(my_ip, None)`) to have all
services bind to their default port. Binding to a specific port is still
supported, but until an API is added to select ports per transport, this
only works well when the AIOCOAP_SERVER_TRANSPORT environment variable
restricts the set of transports -- otherwise, TCP and TLS would try to
bind to the same port.