Factor out hostportsplit
The same idiom has been used in far too many places and was moved into
util.

This incurs a slight change in functionality as the proxy will now
explicitly state all explicitly given coap ports in a Uri-Port option
(and not silencing COAP_PORT), but that's a good thing as the library
now deals with both coap and coaps schemes.