Decoding: raise UnparsableMessage rather than Index/ValueError
Option decoding should raise UnparsableMessage errors because those
propagate through the right paths of "This is not a CoAP message" and
terminate a TCP connection.

This was found when using nmap on a coap+tcp server.

Some error details in the unit tests change; those are OK because they
are more testing whether the implementation recognizes the issue at all
than what a library user can expect from the API.