OSCORE client: be a transport
This change makes the OSCORE client functionality available to
requesters in general; protected requests are executed as soon as a
suitable security context is entered into the credentials.

This has been attempted before but was only realistically possible now
that transports are a lot more flexible and don't necessarily need to
deal with message IDs or even tokens.

It does not fully solve the regression yet because observations are not
fully functional, as that appears to be more of a server-side issue
(both for plugtests 4 and 5, the last message does contain an outer
observe option).

Contributes-To: https://github.com/chrysn/aiocoap/issues/105