observation server-side: Pass around whether the message is the last
This allows observable resources to willingly terminate observations at
any time. The method is a bit convoluted (it involves calling the
otherwise deprecated .deregister() method of the ServerObervation when
generating the first response, and triggering with an additional
parameter at later times), but it is consistent now again with the
earlier behavior, and makes OSCORE observations work properly too.

Closes: https://github.com/chrysn/aiocoap/issues/105