tests: Always show available log output on failed tests
Sharpening the condition on which a test is considered successful (the
previous condition for the execution of the del_to_be_sure checks never
fully worked; the errors array is populated also with successes by the
unittest library).

Originally, log dumps were often displayed because failures triggered
the "context survived destruction" errors. Those were cut down as
annoying in failed cases, but the logs were useful, so they are now
shown independently of the "context survived destruction" cases whenever
a test failed.