setup.py: Make docs not depend on tinydtls
That used to reside in the readthedocs fixes branch, but that it is not
required for docs is true in general.

This makes that branch obsolete again (ie. aiocoap docs can be built
from master again), as by now, no requirements.txt with `-e
git+https://github.com/rtfd/recommonmark.git#egg=recommonmark` is needed
any more to get a recent recommonmark that works with sphinx-argparse.