Move default response code to render_* mechanism
The default response code CONTENT used to be set deep in the protocol
stack; that necessitated some duplication with OSCORE, and was generally
considered a bad idea.

This moves the setting of a default code into the fallback .render()
method of the Resource class, which is also responsible for fanning out
into render_get / render_post etc.

In the course of this change, the default response code now depends on
the request code; GET/FETCH default to Changed, DELETE defaults to
Deleted and everything else to Changed.

This is technically a breaking change; in practice, as most resources
use the render_* methods, they still get default codes; the previous
code of Content was unsuitable for DELETE, PUT and POST anyway.

News-Section: Breaking changes