OSCORE: Use RequestIds to pass around can_reuse_partiv/_nonce
This changes the behavior of observations to respond with the sender's
nonce on the first response. It should make things a bit safer w/rt
future code changes, as the decision of whether a nonce can (still) be
reused is left to a class that will hand it out only once.