Rework NoResponse
The classical way of returning NoResponse from a render method had the
downside that the resulting sentinel could not be passed around through
the various transport layers because it lacks attributes like .remote.

Rather than forcing the concept, the approach was changed to return
actual response objects but set the no_response option on them. The
lower layers of the message processing then look at whether the code and
no_response setting say that it should be held back, and process it
accordingly.

As almost a side effect, the no_response option in requests is now
handled automatically for render_$code methods.

Returning NoResponse inside render_$code methods is still supported for
compatibility reasons, but raises a deprecation warning.

News-Section: Deprecations
Closes: https://github.com/chrysn/aiocoap/issues/109