recvmsg: Rebuild transport based on exported asyncio interfaces
The previous version hooked deep into asyncio internals and was thus
incompatible with uvloop and more recent versions of gbulb. This adds a
more basic transport in its place that does not depend on any of the
internals, but manages a socket on its own and relies on loop.add_reader
to notify it whenever data is pending.

Thanks to Yuri for pointing out this possibility in [1].

[1]: https://groups.google.com/d/msg/python-tulip/06NscaOcL48/nRXOrHFFAAAJ