tokenmanager: Allow sending multicast requests via unicast method
This is not a recommended way of operation, but where the multicast
address is expected to behave rather like an anycast address, it can be
convenient to use the API that will only accept the first response to a
multicast address.

(Such requests previously did not behave well: The responses were not
recognized and answered with a RST if they were CON).