message: Don't set Uri-Port by default
URI parsing used to explicitly set an Uri-Port when a non-standard port
was used, even though when sent to that port, it becomes the default
value of this option.

This update only expects the changed behavior in tests, as the actual
change in behavior is implicit in the upcoming unresolved_remote
changes, but not very visible there.

This does change network observed behavior, but should not break any
applications, for if (and I doubt it) Uri-Port is ever parsed by a
non-proxy server, in any application where a non-default value is meant
to be used, one would need to set it explicitly using .opt.uri_port
anyway.