message: Overhaul storage of request information
The previous content of requested_* and unresolved_remote was a cesspool
of historically developed attributes -- stemming from the wish to
transport the request's details into the response, but ending up being
used for scheme selection as well.

This update splits the information there into what was roughly
.unresolved_remote (currently in ._remote_hints, possibly moving into
.remote), and a single .request attribute that plainly indicates the
full request message that was used to create the reponse.

The old properties are still available as accessors, but due for
deprecation.