message.get_request_uri(): Don't report host name if not transmitted
The expected behavior of get_request_uri() was previously to report the
configured host's name in a response even if that host name was not
transmitted. That behavior was questionable as the server could not
certainly know the name (or more precisely: know whether it was accessed
by IP literal or its name, even if unique), which would result in server
and client disagreeing on the server's name even without reverse proxies
involved.

The expected behavior is now to only express the host name if it was
actually transmitted, otherwise to display the default (typically
literal).

This not only gives more agreed-on results, but also helps in
applications where an initial request is processed by some mechanism
(eg. URI discovery): That mechanism does not need an explicit setting
for whether or not to send the Uri-Host, because if the host is not set
in the first step, URIs built from get_request_uri() will already have a
literal in there.

News-Section: Breaking changes