message: Store remote infos in .remote even for unresolved ones
The unresolved remote information, previously stored in _remote_hints
and before in unresolved_remote and requested_scheme, is now stored in
an EndpointAddress object UndecidedRemote.

As an intended side effect, it is now no longer sufficient to set
uri_host and no uri, reflecting the new reality of there not being a
single CoAP protocol any more, but multiple schemes to choose from.

Access is still available over the old accessors, and a
unresolved_remote can still be set with the effect of setting the
scheme to 'coap'. Applications that relied purely on sending uri_host
may face breakage, though, and are suggested to set a full URI or set
unresolved_remote in addition. (That path is kept as recommended one as
that can be smoothly deprecated later, while mechanisms looking into the
uri_host during the dispatch decision would have a harder time raising
warnings).