options: Don't carry length around
Options used to have a length property.

Given that Python is far from embedded, and that every time an option's
length is queried, it's encoded right after, and the issue of [130],
that practice is abandoned and replaced with querying the encoded
option's length.

A test that covered the encoding lengths of integers was removed, as
that can't be meaningfully executed any more -- the length of the
options is already tested for where their values are compared to
hand-encoded reference values.

[130]: https://github.com/chrysn/aiocoap/issues/130