Correctly increment block cursor when using BERT
BERT requires peers to use block numbers for a block size of 1024, but
allows them to put more than one block in a frame. The current code
always increases the block number by one in BLOCK1 requests, when it
should actually increase by the number of blocks that went out with the
previous frame.

This fixes that.