udp6: Resolve address before binding
Binding to a UDP port requires the zone identifier to be split off the
IP address. Running input host names used for binding through
`getaddrinfo` produces a tuple usable for binding even for those cases.

As a side effect, it is now possible easily to bind to the udp6
transport to an IPv4 address when V6MAPPED is available.

Closes: https://github.com/chrysn/aiocoap/issues/164