remotes: Allow modification for multicast responses
The pktinfo of a returning message was previously stripped from
responses to multicast at MessageInterfaceUDP6.send in a fall-back way.

Transports now have an explicit point to modify their addresse for
responding, which is exercised by the only multicast capable transport udp6.

The warning check was removed as this is now an error (the OS will raise
when trying to send with a multicast source address).