OSCORE: Change the format of stored security contexts
The previous approach of having sharable static files for the common
context that would be open in a "client" and "server" mode proved more
confusing than useful.

Key generation was removed because it was unused, underparametrized (no
way to enter algorithm or even key IDs) and didn't align with the new
model.

News-Section: Breaking changes