udp6: Portability fix to non-Linux platforms
The options required for udp6 were assumed to be portable; turns out
some of them are Linux proprietary.

Consequently, the error handling part (for which there appears to be no
standardized API) are gated behind a platform specific flag, and the
socknumbers module updated accordingly.

Using information gained from a build of a [35569]-fixed Python, numbers
potentially unavailable from the socket module can now fall back to
platform dependent shipped constants. The Python-private `IN` module is
still used, but the fallback definitions are platform-gated.

It is unlikely that this should cause regressions, as the previously
hard-coded constants were Linux specific and are still used on that
platform.

This should allow running the udp6 transport on macOS even without
[35569] fixed, but does not change the default transport selection yet.

[35569]: https://bugs.python.org/issue35569

News-Section: Portability