udp6: Set outgoing address using pktinfo
As the sin6_scope_id of the remote address only has meaning for
link-local addresses (at least in Linux, probably per specification),
requests to ff05::1 can otherwise not be sent.

One small change for the regular cases is that the scope ID of a udp6
endpoint address is not compared any more for equality and hashing. This
is unlikely to cause trouble; a redesign of the UDP6EndpointAddress
that's more oriented towards addresses and less towards socket addresses
(and that would have an interface identifier, and a flag on whether that
was intentionally set or not) might roll that back.

See-Also: https://github.com/chrysn/aiocoap/issues/140