udp6: Use ipaddress standard module
instead of manually checking and hoping that the address is normalized
(which, though, it usually is in that situation as it comes from the
OS).

This also allows for a shortcut through de- and encoding of an address
as IPv6Address accepts binary addesses, and can no-op through its
constructor.