error handlers: Guard error rendering
As this is typically user implemented code, it is expected to fail
easily; in addition (and that's what triggered this), returning None
from to_message is a more-common-than-should-be error (see
aiocoap-fileserver commit following this).

The application bugs prevented by this are otherwise subtle to debug:
The None passed on to the PlumbingRequest causes the request to fail
without a response.