log: Defer formatting to the formatter
Now that the test suite's logger does some initial formatting, objects
can just be left as is in the message without later looking like an
object leak.

Now, if the log message is discarded right away, the cost of formatting
is avoided -- plus it's more readable and more the way it's designed to
be.