aiocoap-client: Relax JSON parsing to Python literals
Not all CBOR can be input as JSON (no numerics as keys, no byte
literals). If a payload string can not be interpreted as JSON, it is
tried as Python. This should create no ambiguities.

(Note that just using Python would force habitual JSON users to
capitalize their constants and rewrite their null)