Proxy: Make composable (including error changes)
This allows proxy components to fall back to each other and to other
site classes.

The special handling of reverse proxies can be removed, as the
raise_unless_safe already handles suprising cases of Proxy-* (the error
messages change, but not really for the worse).

The ReverseProxy class thus becomes quite moot and gets deprecated.