Increase Python dependency version to 3.7
Integration tests for earlier Python versions have already been removed,
and issues like [54] show that earlier versions do show subtle
variations that'd be hard to find in detail.

This does not close the 3.7 transition in full yet (and thus doesn't
close the issue, and leaves the tested-for version at 3.6.9), as PyPy
still doesn't support all of Python 3.7; especially, the code will not
use variable annotations yet.

[54]: https://gitlab.com/oscore/liboscore/-/issues/54

Contributes-To: https://github.com/chrysn/aiocoap/issues/153