Replace exchange_monitor relics with functional messageerror_monitor
The exchange_monitor was a (by now completely unused -- especially, the
application-facing parts were already removed) concept to allow
applications to gather callbacks (primarily for progress reporting /
logging) on individual requests, that fell apart when the transports
were split up into layers.

Especially, it did not work (or was not used) for the one thing where it
would have mattered in token-layer to message-layer communication: to
tell the token layer that one of the messages it sent was actively
rejected. (That matters a only little in the request case where the RST stopped
retransmissions and eventually timed out by the application, because
RSTs to requests are rare, and doesn't matter at all in regular
responses where either an ACK or a RST are good to stop the CONs from
flowing, but matters a lot in observation CONs where the notifications
would keep flowing even though a RST was received).

Therefore, it was replaced with a smaller mechanism to take that place
that does exactly that.