udp6: Show warning when no pktinfo is available
As the setsockopt is required to succeed before a udp6 is ready, it
stands to reason to expect a pktinfo with every message.

This is only a warning and no error as even when this breaks client mode
does work -- but there will be otherwise hard to understand errors when
a request comes in. The warning is supposed to help making the issue
more visible by raising it early on, possibly even in test cases (though
if on a host no pktinfo is sent at all, the server tests should already
fail).

Contributes-To: https://github.com/chrysn/aiocoap/issues/209