OSCORE: Remove leftover mechanism from old OSCOAP
Back in early draft versions, the request nonce could only be used in
unicast mode as it was just about flipping a bit. Now that the sender ID
of the party whose sequence number it is is part of the nonce, all
responses to multicast can work without an own sequence number.