OSCORE: Generalize verify_start and context discovery
With original OSCORE, only the KID and KID context influenced the
decision of which context to pick; with group OSCORE, there needs to be
more information (like "Is group mode set? If not, pick pairwise mode").

Thus, start_verify now just exposes the whole pack of header data, which
can also be passed on to the available contexts so they can declare
whether they feel responsible.