DTLS: Overhaul error handling, ensuring dead connections are not reused
The previous handling was inconsistent in that some errors resulted in
an attempt to reuse the connection object (which is not in line with the
planned use of remotes, cf. [225]).

Instead, DTLS remotes now remove themselves from the pool as soon as
they can if anything troublesome crops up, and more reliably report the
errors.

The internal parameter connect_immediately was removed because now there
is just no valid remote until someone creates one anew by requesting the
address from the (now empty) pool.

[225]: https://github.com/chrysn/aiocoap/issues/225