Return -1 when weak reference does not exist in the wrapper for accessor function.
At the moment null is returned when reference does not exist, which
causes issues for `_write` and `_read` functions in `DTLSSocket.dtls`
module. For `_event` function, it does not matter at the moment, it
seems.

But all three functions in the `DTLSSocket.dtls` module return an
integer, so I would expect that's the contract for all of them.

Add some type annotations, which before this pull request would point
the issue

    $ mypy aiocoap | grep tiny
    aiocoap/transports/tinydtls.py:167: error: Return value expected

I can remove the type annotations if the type hints are a no go for this
project.