RD: Add basic support for security policies
This allows users of the resource directory implementation to define
policies policies, i.e. limit access to specific endpoint names and sectors.

The configuration happens using a separate JSON file that needs to be
provided with the `--security-policy` command line option.

The file format is currently described in the rd.py module description, a
CDDL description (like in credentials.cddl) is not included yet.