numbers: Replace media_type{,_rev} with ContentFormat
The old mechanism was crude (in particular: accept and content_format
needed manual lookups by the user), and incomplete (missing the encoding
column)

The old lookups are deprecated, with compatibility maintained including
old workarounds (that now don't need a dedicated deprecation warning
because the whole mechanism was deprecated).

Including a fix to extensible enums to not trip over class methods.