resource interface: Allow rendering to plumbing request
This bypasses a lot of established mechanisms, which are to be migrated
to wrappers on this interface. As they are being migrated, more and more
resources can retroactively automatically implement
render_to_plumbingrequest, and OSCORE can do blockwise and observations
by accessing the innter site through its render_to_plumbingrequest
rather than the other interfaces.

Contributes-To: https://github.com/chrysn/aiocoap/issues/251