tests: Remove odd reference cycle that doesn't seem to match up with the current setup any more
No hint was found at the introduction time (across moves) as to why that
loop needs to be present to prove any better behavior. It is admittedly
odd that that very test would start failing with the most recent tests;
still treating it as a GC regression at an irrelevant test because
the test's documentation failed to show its relevance.