Copy over LinkHeader library
The library has not been actively maintained for several years, and the
lack of a py3-none-any.whl is impeding usage with pyodide and Jupyter.

The file was copied from version 0.4.3 as on PyPI, with a SPDX header
added (with data from its setup.py). The file is otherwise unmodified,
and exempt from PEP8 checks.