CI: Use python:3 instead of Debian image for assembling pages
Debian images need extra steps just to get pip, and on top of that
started complaining about pip packages being installed. (Installing
through pip rather than dpkg was started because the reporting coverage
needs to be as new as the others, and given we're testing on various
Debian and non-Debian images, PyPI is the best source for a current
package).