udp6: Emulate own better AI_ADDRCONFIG (take II)
The previous approach was flawed because AI_ADDRCONFIG doesn't test
reachability, it's just a very coarse filter. Attempting to connect a
UDP socket explicitly asks the OS.

This is not the solution to all IPv6 worries one could have (one could
have a route but it's not serviced), but that has clear fault with the
network provider, and working around that would require assessments such
as "was that request safe to send again", and fallbacks in case the
first request is a POST that, under happy-eyeballs logic, might be
processed twice.