credentials: Rename contextfile to basedir for oscore
The old name is deprecated. The __init__ signature gains type
annotations so that when the deprecation is removed, the full workaround
can just be removed. Some mypy fixes ensued.

This makes use of the recent factory-ification of the type map; while
not being a reduction complexity on its own, it will be once the
deprecation is resolved.